+++ src/constants.cpp
@@ -23,7 +23,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include <wx/string.h>
 #include <wx/filefn.h>
 #include <curl/curl.h>
-#include "rapidjson/rapidjson.h"
+#include <rapidjson/rapidjson.h>
 #include "db/DB_Upgrade.h"
 
 /*************************************************************************
