+++ src/db/DB_Table.h
@@ -27,13 +27,13 @@
 #include <wx/wxsqlite3.h>
 #include <wx/intl.h>
 
-#include "rapidjson/document.h"
-#include "rapidjson/pointer.h"
-#include "rapidjson/prettywriter.h"
-#include "rapidjson/stringbuffer.h"
+#include <rapidjson/document.h>
+#include <rapidjson/pointer.h>
+#include <rapidjson/prettywriter.h>
+#include <rapidjson/stringbuffer.h>
 using namespace rapidjson;
 
-#include "html_template.h"
+#include <html_template.h>
 using namespace tmpl;
 
 typedef wxLongLong int64;
