+++ src/filtertrans.cpp
@@ -21,7 +21,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "filtertrans.h"
 
 #include "attachmentdialog.h"
-#include "../reports/htmlbuilder.h"
+#include "reports/htmlbuilder.h"
 
 mmFilterTransactions::mmFilterTransactions()
 {
