+++ src/import_export/payeematchandmerge.cpp
@@ -19,7 +19,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 ********************************************************/
 
 #include "payeematchandmerge.h"
-#include "rapidjson/document.h"
+#include <rapidjson/document.h>
 #include <algorithm>
 #include <wx/regex.h>
