+++ src/mmSimpleDialogs.cpp
@@ -19,9 +19,9 @@ along with this program; if not, write to the Free Software
 Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 ********************************************************/
 
-#include "wx/dcbuffer.h"
-#include "wx/popupwin.h"
-#include "wx/spinctrl.h"
+#include <wx/dcbuffer.h>
+#include <wx/popupwin.h>
+#include <wx/spinctrl.h>
 #include <wx/graphics.h>
 #include <wx/renderer.h>
 #include <wx/richtooltip.h>
