+++ src/mmcustomdata.h
@@ -20,8 +20,8 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
 #pragma once
 #include "defs.h"
-#include "Model_CustomField.h"
-#include "Model_Checking.h"
+#include "model/Model_CustomField.h"
+#include "model/Model_Checking.h"
 
 
 class mmDatePickerCtrl;
