+++ src/optionsettingstransaction.h
@@ -21,7 +21,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
 #include "optionsettingsbase.h"
 #include "reports/mmDateRange.h"
-#include "wx/spinctrl.h"
+#include <wx/spinctrl.h>
 
 class OptionSettingsTransaction : public OptionSettingsBase
 {
