+++ src/reports/budgetcategorysummary.cpp
@@ -22,13 +22,13 @@
   Renamed after extensive modifications to original file reportbudgetsetup.cpp
  **************************************************************************/
 #include "budgetcategorysummary.h"
-#include "reports/htmlbuilder.h"
+#include "htmlbuilder.h"
 #include "mmex.h"
 #include "mmframe.h"
 #include "model/Model_Budgetyear.h"
 #include "model/Model_Budget.h"
 #include "model/Model_Category.h"
-#include "reports/mmDateRange.h"
+#include "mmDateRange.h"
 
 mmReportBudgetCategorySummary::mmReportBudgetCategorySummary()
 {
