+++ src/reports/budgetingperf.cpp
@@ -19,13 +19,13 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 ********************************************************/
 
 #include "budgetingperf.h"
-#include "reports/htmlbuilder.h"
+#include "htmlbuilder.h"
 #include "mmex.h"
 #include "mmframe.h"
 #include "model/Model_Budgetyear.h"
 #include "model/Model_Budget.h"
 #include "model/Model_Category.h"
-#include "reports/mmDateRange.h"
+#include "mmDateRange.h"
 #include <string>
 
 mmReportBudgetingPerformance::mmReportBudgetingPerformance()
