+++ src/reports/cashflow.cpp
@@ -22,7 +22,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "mmex.h"
 #include "mmframe.h"
 #include "util.h"
-#include "reports/htmlbuilder.h"
+#include "htmlbuilder.h"
 #include "model/Model_Account.h"
 #include "model/Model_Billsdeposits.h"
 #include "model/Model_CurrencyHistory.h"
