+++ src/reports/categexp.cpp
@@ -21,8 +21,8 @@
 #include "categexp.h"
 #include "budget.h"
 #include "images_list.h"
-#include "reports/htmlbuilder.h"
-#include "reports/mmDateRange.h"
+#include "htmlbuilder.h"
+#include "mmDateRange.h"
 #include "option.h"
 #include <algorithm>
 #include "model/Model_Category.h"
