+++ src/reports/htmlbuilder.h
@@ -24,7 +24,7 @@
 #include "defs.h"
 #include <vector>
 #include "model/Model_Currency.h"
-#include "html_template.h"
+#include <html_template.h>
 #include "util.h"
 
 struct GraphSeries
