+++ src/reports/incexpenses.cpp
@@ -21,9 +21,9 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "images_list.h"
 #include "incexpenses.h"
 
-#include "reports/htmlbuilder.h"
+#include "htmlbuilder.h"
 #include "util.h"
-#include "reports/mmDateRange.h"
+#include "mmDateRange.h"
 
 #include "model/Model_Account.h"
 #include "model/Model_Checking.h"
