+++ src/transdialog.h
@@ -28,10 +28,10 @@
 #include "mmSimpleDialogs.h"
 #include "fusedtransaction.h"
 
-#include "Model_Checking.h"
-#include "Model_Payee.h"
-#include "Model_CustomFieldData.h"
-#include "Model_Splittransaction.h"
+#include "model/Model_Checking.h"
+#include "model/Model_Payee.h"
+#include "model/Model_CustomFieldData.h"
+#include "model/Model_Splittransaction.h"
 
 class mmTextCtrl;
 class wxString;
