+++ src/unix/platfdep.cpp
@@ -17,7 +17,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  ********************************************************/
 
 //----------------------------------------------------------------------------
-#include "../platfdep.h"
+#include "platfdep.h"
 #include <wx/stdpaths.h>
 #include <wx/filename.h>
 //----------------------------------------------------------------------------
